+++ launchdarkly/resource_launchdarkly_feature_flag_test.go
@@ -116,7 +116,7 @@ resource "launchdarkly_feature_flag" "json" {
 	// not support creating members with the same email address more than once.
 	testAccFeatureFlagWithMaintainer = `
 resource "launchdarkly_team_member" "test" {
-	email = "%s@example.com"
+	email = "%s+wbteste2e@launchdarkly.com"
 	first_name = "first"
 	last_name = "last"
 	role = "admin"
@@ -136,7 +136,7 @@ resource "launchdarkly_feature_flag" "maintained" {
 	// the previous maintainer if that maintainer still exists
 	testAccFeatureFlagMaintainerComputed = `
 resource "launchdarkly_team_member" "test" {
-	email = "%s@example.com"
+	email = "%s+wbteste2e@launchdarkly.com"
 	first_name = "first"
 	last_name = "last"
 	role = "admin"